pyproject.toml
[project]
name = "newsfeed-platform"
version = "0.1.0"
description = ""
authors = [
    {name = "Zakaria Jaadi",email = "mr.zakaria.jaadi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic (>=2.11.7,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "streamlit (>=1.48.0,<2.0.0)",
    "plotly (>=6.2.0,<7.0.0)",
    "feedparser (>=6.0.11,<7.0.0)",
    "aiohttp (>=3.12.15,<4.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
